refactor: extract broker runtime event handlers
