+++ .trajectories/index.json
@@ -1,6 +1,6 @@
 {
   "version": 1,
-  "lastUpdated": "2026-05-19T01:38:29.338Z",
+  "lastUpdated": "2026-05-19T01:50:40.535Z",
   "trajectories": {
     "traj_05xg7j388bc4": {
       "title": "Add browser workflow step integration",
@@ -967,6 +967,13 @@
       "startedAt": "2026-05-19T01:28:35.746Z",
       "completedAt": "2026-05-19T01:38:29.105Z",
       "path": "/Users/will/Projects/AgentWorkforce/relay/.trajectories/completed/2026-05/traj_47akjihewlow.json"
+    },
+    "traj_x37bhga2j5ph": {
+      "title": "Deepen broker runtime refactor for PR 906",
+      "status": "completed",
+      "startedAt": "2026-05-19T01:42:10.602Z",
+      "completedAt": "2026-05-19T01:50:40.359Z",
+      "path": "/Users/will/Projects/AgentWorkforce/relay/.trajectories/completed/2026-05/traj_x37bhga2j5ph.json"
     }
   }
 }
